requirements.txt
Django
djangorestframework
AWSIoTPythonSDK